add additional asserts for blank string validation
